Reduce JSONL exporter churn
